Fix Curve ROM for 64bit chunk

+++ version3/java/ROM_BRAINPOOL_64.java
@@ -32,7 +32,7 @@ public class ROM
 	public static final long MConst= 0xA75590CEFD89B9L;
 
 	public static final int CURVE_Cof_I= 1;
-	public static final int[] CURVE_Cof= {0x1L,0x0L,0x0L,0x0L,0x0L};
+	public static final long[] CURVE_Cof= {0x1L,0x0L,0x0L,0x0L,0x0L};
 	public static final int CURVE_A= -3;
 	public static final int CURVE_B_I= 0;
 	public static final long[] CURVE_B= {0xE58101FEE92B04L,0xEBC4AF2F49256AL,0x733D0B76B7BF93L,0x30D84EA4FE66A7L,0x662C61C4L};
